Fix lint

+++ pkg/local/config.go
@@ -18,7 +18,7 @@ type SyncFlags struct {
 type Config struct {
 	SyncFlags
 	// MaxDownloadRetries - Maximum number of retries for download operations
-	MaxDownloadRetries int
+	MaxDownloadRetries uint64
 	// SkipNonRegularFiles - By default lakectl local fails if local directory contains irregular files. When set, lakectl will skip these files instead.
 	SkipNonRegularFiles bool
 	// IncludePerm - Experimental: preserve Unix file permissions
